Simplify url scheme
